django_story_platform/requirements.txt
"""
Django Story Platform - 项目依赖
"""
# Django核心
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# 数据库
psycopg2-binary==2.9.7

# 异步任务
celery==5.3.4
redis==5.0.1

# WebSocket支持
channels==4.0.0
channels-redis==4.1.0

# AI和机器学习
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
diffusers==0.21.4
accelerate==0.24.1

# 图像处理
Pillow==10.1.0
opencv-python==4.8.1.78

# 音频处理
librosa==0.10.1
soundfile==0.12.1

# 视频处理
moviepy==1.0.3

# HTTP客户端
requests==2.31.0
httpx==0.25.2

# 工具库
numpy==1.24.3
pandas==2.1.3
pathlib==1.0.1
python-dotenv==1.0.0

# 开发工具
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
